refactor(review-hub): extract pure helpers and presentational components
- add sumChurn and applySortChange to reviewHubUtils.ts, dedupe the 3x churn reduce and 2x sort-toggle logic
- extract BaseBranchFileRow.tsx (self-contained row component)
- extract PushErrorBanner.tsx (push-error IIFE, state stays parent-owned)
- extract PrStatusChip.tsx (PR-status chip + no-PR IIFE)
